update ship functions to abstract findClosestBounceWall
