+++ src/core/ship/physics/checkForBounce.ts
@@ -4,21 +4,19 @@
  */
 
 import type { Store } from '@reduxjs/toolkit'
-import type { MonochromeBitmap, LineRec } from '@core/walls'
+import type { MonochromeBitmap } from '@core/walls'
 import type { ShipState } from '@core/ship'
-import type { Point } from '@core/shared/pt2xy'
 import { blackTerrain } from '@core/walls/render/blackTerrain'
 import { checkFigure, shipSlice } from '@core/ship'
 import { eraseFigure } from '@core/ship/render'
 import { SCENTER } from '@core/figs'
 import { LINE_KIND } from '@core/shared'
-import { pt2line } from '@core/shared/pt2line'
-import { getstrafedir } from '@core/shared/getstrafedir'
+import {
+  findClosestBounceWall,
+  type WallData as CheckForBounceData
+} from './findClosestBounceWall'
 
-export type CheckForBounceData = {
-  kindPointers: Record<number, string | null>
-  organizedWalls: Record<string, LineRec>
-}
+export type { CheckForBounceData }
 
 export type CheckForBounceDeps = {
   screen: MonochromeBitmap
@@ -114,53 +112,4 @@ export function checkForBounce(deps: CheckForBounceDeps): MonochromeBitmap {
     // Return the screen with bounce walls rendered (ship not erased)
     return screenWithBounceWalls
   }
-}
-
-/**
- * Find the closest bounce wall to the ship and return the norm direction.
- * Corresponds to bounce_ship() at Play.c:302-313 and getstrafedir() at Terrain.c:242-263
- */
-function findClosestBounceWall(
-  globalx: number,
-  globaly: number,
-  unbouncex: number,
-  unbouncey: number,
-  wallData: CheckForBounceData,
-  _worldwidth: number
-): { norm: number } | null {
-  const firstBounceId = wallData.kindPointers[LINE_KIND.BOUNCE]
-  if (!firstBounceId) return null
-
-  let closestWall: LineRec | null = null
-  let minDistance = 1000 // Start with large squared distance like original (Play.c:304)
-
-  // Create Point object for ship position (Play.c:302-303)
-  const shipPoint: Point = { h: globalx, v: globaly }
-
-  // Check all bounce walls (Play.c:305-310)
-  let lineId: string | null = firstBounceId
-  while (lineId !== null) {
-    const line: LineRec | undefined = wallData.organizedWalls[lineId]
-    if (!line) break
-
-    // Calculate actual distance from ship to line segment using pt2line
-    // This matches the original pt2line() call at Play.c:306
-    const distance = pt2line(shipPoint, line)
-
-    // Compare squared distances (pt2line returns squared distance)
-    if (distance < minDistance) {
-      minDistance = distance
-      closestWall = line
-    }
-
-    lineId = line.nextId
-  }
-
-  if (!closestWall) return null
-
-  // Calculate norm using getstrafedir (Play.c:313)
-  // The norm should point away from the wall toward unbouncex/unbouncey
-  const norm = getstrafedir(closestWall, unbouncex, unbouncey)
-
-  return { norm }
 }
